Revert "Update clim_adjust_common_input.py" see issue #130
This reverts commit 45d26d7d4f32211df88c2fc49863cc2951ae27ef.

It should be implemented at a different level. One problem is that
info/clim_adjust_common_input.py is not supposed to access runner data,
because it's common input for all countries. If one day we decide to
implement weighting by the inventory area, we would need to put this
code into info/clim_adjust, which contains the country specific
information and scenario combo information with the inventory area